pyproject.toml
[tool.poetry]
name = "agents"
version = "0.1.0"
description = ""
authors = ["Joel Pires <joelpires70@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.8.1"
python-dotenv = "^1.0.1"
pymongo = "^4.9.1"
mongoengine = "^0.29.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pymongo-inmemory = "^0.4.2"
mongomock = "^4.1.2"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
